refactor: migrate features to src dir
