+++ src/features/authorization/EmailSignupForm.tsx
@@ -1,12 +1,12 @@
 import { useForm } from 'react-hook-form';
 import { Alert, Button } from '@mui/material';
-import { PasswordInput } from '../../src/components/textfield/PasswordInput';
+import { PasswordInput } from '../../components/textfield/PasswordInput';
 import { useRouter } from 'next/router';
 import { SignupPostResolver } from '@/schemas/SignupSchema';
 import { mutate } from 'swr';
 import { useState } from 'react';
 import { BrothTextField } from '@/components/textfield/BrothTextField';
-import { swaggerClient } from '../../src/init/swaggerClient';
+import { swaggerClient } from '../../init/swaggerClient';
 import { User } from '@/init/swagger';
 import { Stack } from '@mui/system';
